requirements.txt
beautifulsoup4==4.10.0
ccy==1.2.2
coinpaprika==0.1.0
dateparser==1.0.0
Django==3.2.7
drf_yasg==1.20.0
firebase_admin==5.2.0
forex_python==1.6
GoogleNews==1.6.0
newspaper3k==0.2.8
nltk==3.6.3
pandas==1.3.3
pycoingecko==2.2.0
pymongo==3.12.0
readtime==1.1.1
regex==2021.9.24
requests==2.26.0
celery~=5.1.2
djangorestframework~=3.13.1
APScheduler~=3.8.1